- Add react router - code refactor   - separation of components
