+++ package.json
@@ -12,6 +12,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-infinite-scroll-component": "^5.0.4",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
     "react-scroll": "^1.7.16",
     "semantic-ui-css": "^2.4.1",
@@ -20,7 +21,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "demo": "export REACT_APP_API_ROOT=http://api.animalcrossingart.test:1337 && react-scripts start",
+    "demo": "export REACT_APP_API_ROOT=http://api.animalcrossingart.test:1337 PORT=3009 && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
