+++ src/Containers/Home.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Gallery from '../Components/Gallery/Gallery';
+
+const Home = () => (
+  <div>
+    <Gallery />
+  </div>
+);
+
+export default Home;
